Qt Designer: Fix range check of double properties.

+++ src/shared/qtpropertybrowser/qtpropertymanager.cpp
@@ -880,7 +880,7 @@ public:
 
     struct Data
     {
-        Data() : val(0), minVal(-INT_MAX), maxVal(INT_MAX), singleStep(1), decimals(2) {}
+        Data() : val(0), minVal(-DBL_MAX), maxVal(DBL_MAX), singleStep(1), decimals(2) {}
         double val;
         double minVal;
         double maxVal;
